pyproject.toml
[tool.poetry]
name = "tvpy"
version = "0.0.7"
description = "📺 TvPy"
homepage = "https://github.com/gkutiel/tvpy/tree/master"
repository = "https://github.com/gkutiel/tvpy/tree/master"
documentation = "https://github.com/gkutiel/tvpy/tree/master"
readme = "README.md"
authors = ["Gilad Kutiel <gilad.kutiel@gmail.com>"]
include = ['tvpy/*']

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
fire = "^0.4.0"
tqdm = "^4.64.0"
Pillow = "^9.2.0"
websockets = "^10.3"
rich = "^12.5.1"
parse-torrent-title = "^2.4"
pyright = "^1.1.273"
"fs.smbfs" = "^1.0.5"
toml = "^0.10.2"
1337x = "^1.2.3"
libtorrent = "^2.0.7"

[tool.poetry.dev-dependencies]
rope = "^1.3.0"
autopep8 = "^1.7.0"
pytest = "^7.1.2"
pytest-watch = "^4.2.0"
doit = "^0.36.0"
isort = "^5.10.1"
autoflake = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tv-json = "tvpy.main:tv_json"
tv-info = "tvpy.main:tv_info"

tv-renm = "tvpy.main:tv_renm"
tv-sync = "tvpy.main:tv_sync"
tv-html = "tvpy.main:tv_html"
